Unify method implementation, IsEmpty and Clear operations in order for the union find to implement an Abstract Data Type

+++ unionfind/constructor.go
@@ -3,7 +3,7 @@ package unionfind
 func Empty() *UnionFind {
 	return &UnionFind{
 		size: 0,
-		parents: make([]int64, 0, 0),
+		parents: make([]int, 0, 0),
 		setSizes: make([]int64, 0, 0),
 	}
 }
@@ -16,10 +16,10 @@ func New(size int64) *UnionFind {
 	u := &UnionFind{}
 	u.setSize(size)
 	u.setNumSets(size) // each element is a component in the beginning, so size = numSets
-	u.parents = make([]int64, size)
+	u.parents = make([]int, size)
 	u.setSizes = make([]int64, size)
 	for i := int64(0); i < size; i++ {
-		u.parents[i] = i
+		u.parents[i] = int(i)
 		u.setSizes[i] = 1
 	}
 	return u
